Make ie8 asset output filename unique
Fixes asset build error where assets clean will remove the source file
with the same name as the output file.

The steps that caused the issue are;

1. run collect static
- ie8.css copied from cmsbootstrap to project STATIC_ROOT

2. run assets clean
- assets removes the ie8.css input file as it has the same name as the
  bundle output file.

3. run assets build
- assets unable to find the input file to rebuild the bundle.

When ASSETS_DEBUG is False django-assets won't use static finders to
locate the input file for the bundle - it instead expects to find ALL
files in the project STATIC_ROOT directory (this issue was not triggered
when debug is True, as only then django-assets runs static finders
itself).